+++ OnlySnarf/classes/message.py
@@ -8,7 +8,7 @@ from .poll import Poll
 from .schedule import Schedule
 from ..util.defaults import PRICE_MAXIMUM, PRICE_MINIMUM
 from ..util.settings import Settings
-from ..util.webdriver import message, post
+from ..util.webdriver import message as WEBDRIVER_message, post as WEBDRIVER_post
 
 from marshmallow import Schema, fields, validate, post_load
 
@@ -37,14 +37,13 @@ class Message():
 
         """
 
-        self.files = files
+        self.files = Message.format_files(files)
         self.keywords = keywords
         self.performers = performers
-        self.price = price
-        self.recipients = recipients
-        self.schedule = schedule        
-        self.text = text
-        self.__initialized__ = False
+        self.price = Message.format_price(price)
+        self.recipients = list(set(recipients))
+        self.schedule = Message.format_schedule(schedule)        
+        self.text = Message.format_text(text, keywords, performers, files)
 
     @staticmethod
     def create_message(message_data):
@@ -56,17 +55,6 @@ class Message():
         result = schema.dump(self)
         # pprint(result, indent=2)
         return result
-
-    def init(self):
-        """Initialize."""
-
-        if self.__initialized__: return
-        self.get_text()
-        self.get_keywords()
-        self.get_price()
-        self.get_files()
-        self.get_performers()
-        self.__initialized__ = True
 
     @staticmethod
     def format_keywords(keywords):
@@ -84,7 +72,7 @@ class Message():
             The generated keywords into a string
         """
 
-        # ternary: a if condition else b
+        keywords = [n.strip() for n in keywords]
         return "#{}".format(" #".join(keywords)) if len(keywords) > 0 else ""
 
     @staticmethod
@@ -104,11 +92,11 @@ class Message():
             The generated performers into a string
         """
 
-        # ternary: a if condition else b
+        performers = [n.strip() for n in performers]
         return " @{} ".format(" @".join(performers)) if len(performers) > 0 else ""
 
-
-    def format_text(self):
+    @staticmethod
+    def format_text(text, keywords=[], performers=[], files=[]):
         """Formats self.text with the provided keywords and performers
 
         
@@ -120,8 +108,82 @@ class Message():
 
         """
 
-        return "{}{}{}".format(self.get_text(), Message.format_performers(self.get_performers()), Message.format_keywords(self.get_keywords())).strip()
+        if not text and len(keywords) == 0 and len(performers) == 0 and len(files) == 0:
+            Settings.warn_print("formatting empty message!")
+            return ""
+        if not text and len(files) > 0:
+            text = self.get_text_from_filename(files[0])
+        return f"{text}{Message.format_performers(performers)}{Message.format_keywords(keywords)}".strip()
 
+    @staticmethod
+    def format_schedule(**scheduleArgs):
+        return Schedule(**scheduleArgs).dump()
+
+    # TODO: reintegrate upload max
+    @staticmethod
+    def format_files(file_paths):
+        files = []
+        if len(file_paths) > 0:
+            # for file in files[:int(Settings.get_upload_max())]:
+            for file in file_paths:
+                if not isinstance(file, File):
+                    files.append(File(file))
+                else:
+                    files.append(file)
+        # files = files[:int(Settings.get_upload_max())] # reduce by max
+        return files
+
+    @staticmethod
+    def format_price(price):
+        if not price or str(price) == "0": return 0
+        if str(price) == "max": return Decimal(PRICE_MAXIMUM)
+        elif str(price) == "min": return Decimal(PRICE_MINIMUM)
+        elif Decimal(sub(r'[^\d.]', '', str(price))) < Decimal(PRICE_MINIMUM):
+            Settings.warn_print(f"price too low: {price} < {PRICE_MINIMUM}")
+            Settings.maybe_print("adjusting price to minimum...")
+            return Decimal(PRICE_MINIMUM)
+        elif Decimal(sub(r'[^\d.]', '', str(price))) > Decimal(PRICE_MAXIMUM):
+            Settings.warn_print(f"price too high: {price} < {PRICE_MAXIMUM}")
+            Settings.maybe_print("adjusting price to maximum...")
+            return Decimal(PRICE_MAXIMUM)    
+        return Decimal(price)
+
+    # TODO: format filename better, possibly grab text from parent folder instead
+    def get_text_from_filename(self, file_object):
+        """Gets text from this object's file's title"""
+
+        if not self.get_files(): return ""
+        text = file_object.get_title()
+
+        return text.replace("_", " ")
+
+    # TODO: add updates to user object upon successful message?
+    def on_success(self):
+        pass
+
+    def send(self):
+        """
+        Sends this message.
+
+
+        Returns
+        -------
+        bool
+            Whether or not sending the message was successful.
+
+        """
+
+        try:
+            return WEBDRIVER_message(self.dump())        
+        finally:
+            self.on_success()
+
+
+
+
+
+
+    # TODO: probably move this to webdriver equivalent
     @staticmethod
     def is_tip(text):
         """
@@ -151,221 +213,12 @@ class Message():
             return True, int(amount)
         return False, 0
 
-    def get_files(self):
-        """
-        Gets files from args specified source or prompts as necessary.
 
-        Uses appropriate file select method as specified by runtime args:
-        - remote (server, ipfs)
-        - local
-
-        Parameters
-        ----------
-        again : bool
-            Whether or not it is the script user's first time around.
-
-        Returns
-        -------
-        list
-            Files in a list
-
-        """
-
-        if len(self.files) > 0:
-            files_ = []
-            for file in self.files[:int(Settings.get_upload_max())]:
-                if not isinstance(file, File):
-                    file_ = File()
-                    setattr(file_, "path", file)
-                    files_.append(file_)
-                else:
-                    files_.append(file)
-            return files_
-            # return self.files[:int(Settings.get_upload_max())]
-        files = Settings.get_input_as_files()
-        if len(files) > 0:
-            Settings.dev_print("fetched input files for upload")
-            self.files = files[:int(Settings.get_upload_max())] # reduce by max
-            # self.files = files
-            # return files
-        return self.files
-        # files = Folder.get_files()
-        # if files is empty this all basically just skips to the end and returns blank 
-        # filed = []
-        # for file in files:
-            # turn all folders into their files
-            # if isinstance(file, Folder): filed.extend(file.get_files())
-            # else:
-                # filed.append(file)
-                # TODO
-                # this goes elsewhere
-                # flag that the files include a performer
-                # if hasattr(file, "performer"):
-                    # self.performers.append(getattr("performer", file))
-
-    def get_message(self):
-        """
-        Gets the message as a serialized JSON object.
-
-
-        Returns
-        -------
-        Object
-            The message as an object.
-
-        """
-
-        return dict({
-            "text": self.format_text(),
-            "files": self.get_files(),
-            "performers": self.get_performers(),
-            "price": self.get_price(),
-            "keywords": self.get_keywords()
-        })
-
-    def get_performers(self):
-        """
-        Gets the performers for the text.
-
-        Returns
-        -------
-        list
-            The performers
-
-        """
-
-        if len(self.performers) > 0: return self.performers
-        self.performers = Settings.get_performers()
-        return self.performers
-
-    def get_price(self):
-        """
-        Gets the price value if not none else sets it from args or prompts.
-
-
-        Returns
-        -------
-        int
-            The price
-
-        """
-
-        if self.price: return self.price
-        price = Settings.get_price()
-        if str(price) == "0": return 0
-        priceMin = Settings.get_price_minimum()
-        priceMax = Settings.get_price_maximum()
-        if str(price) == "max": price = priceMax
-        elif str(price) == "min": price = priceMin
-        elif Decimal(sub(r'[^\d.]', '', str(price))) < Decimal(priceMin):
-            Settings.warn_print("price too low: {} < {}".format(price, priceMin))
-            Settings.maybe_print("adjusting price to minimum...")
-            price = priceMin
-        elif Decimal(sub(r'[^\d.]', '', str(price))) > Decimal(priceMax):
-            Settings.warn_print("price too high: {} < {}".format(price, priceMax))
-            Settings.maybe_print("adjusting price to maximum...")
-            price = priceMax    
-        self.price = price
-        return self.price
-
-    def get_keywords(self):
-        """
-        Gets the keywords for the text.
-
-        Returns
-        -------
-        list
-            The keywords
-
-        """
-
-        if len(self.keywords) > 0: return self.keywords
-        self.keywords = Settings.get_keywords()
-        return self.keywords
-
-    # cannot have duplicates
-    def get_recipients(self):
-        self.recipients = list(set(self.recipients))
-        return self.recipients
-        # if len(self.recipients) > 0: 
-
-    def get_text(self, again=True):
-        """
-        Gets the text value if not none else sets it from args or prompts.
-
-
-        Parameters
-        ----------
-        again : bool
-            Whether or not it is the script user's first time around.
-
-        Returns
-        -------
-        str
-            The text to enter.
-
-        """
-
-        if self.text != "": return self.text
-        # retrieve from args and return if exists
-        text = Settings.get_text()
-        if text != "": 
-            self.text = text
-            return text
-        text = self.get_text_from_filename()
-        if text != "":
-            self.text = text
-            return text
-        self.text = text
-        return self.text
-
-    def get_text_from_filename(self):
-        """Gets text from this object's file's title"""
-
-        if not self.get_files(): return ""
-        text = self.files[0].get_title()
-        # if "_" in str(self.text):
-        if re.match("[0-9]_[0-9]", text) is not None:
-            texttext = self.files[0].get_parent()["title"]
-        else:
-            try: 
-                int(text)
-                # is a simple int
-                if int(text) > 20:
-                    text = self.files[0].get_parent()["title"]
-            except Exception as e:
-                # not a simple int
-                # do nothing cause probably set already
-                pass
-        text = text.replace("_", " ")
-        # redo keyword parsing (unsure if necessary call)
-        text = self.update_keywords(text)
-        return text
-
-    def on_success(self):
-
-    def send(self):
-        """
-        Sends this message.
-
-
-        Returns
-        -------
-        bool
-            Whether or not sending the message was successful.
-
-        """
-
-        try:
-            self.init()
-            return message(self.get_message())        
-        finally:
-            self.on_success()
 
 class Post(Message):
     """OnlyFans message (and post) class"""
 
-    def __init__(self):
+    def __init__(self, expiration, poll, **kwargs):
         """
         OnlyFans post object
 
@@ -374,108 +227,13 @@ class Post(Message):
 
         """
 
-        super().__init__(self)
-        self.expiration = 0
-        self.poll = None
-
-    # def __str__(self):
-    #     return "fooPost"
-
-    def init(self):
-        """Initialize."""
-
-        super().init()
-        self.__initialized__ = False
-        self.get_poll()
-        self.get_schedule()
-        self.get_expiration()
-        self.__initialized__ = True
-
-    def get_expiration(self, again=True):
-        """
-        Gets the expiration value if not none else sets it from args or prompts.
-        
-        Parameters
-        ----------
-        again : bool
-            Whether or not it is the script user's first time around.
-
-
-        Returns
-        -------
-        int
-            The expiration as an int.
-
-        """
-
-        if self.expiration: return self.expiration
-        # retrieve from args and return if exists
-        expiration = Settings.get_expiration() or 0
-        if expiration: 
-            self.expiration = expiration
-            return expiration
+        super().__init__(self, **kwargs)
         self.expiration = expiration
-        return self.expiration
+        self.poll = Post.format_poll(poll)
 
-    def get_poll(self, again=True):
-        """
-        Gets the poll value if not none else sets it from args or prompts.
-        
-        Parameters
-        ----------
-        again : bool
-            Whether or not it is the script user's first time around.
-
-
-        Returns
-        -------
-        Poll
-            Poll object with proper values
-
-        """
-
-        # check if poll is ready
-        if self.poll: return self.poll
-        self.poll = Poll()
-        return self.poll
-
-    def get_post(self):
-        """
-        Gets the message as a serialized JSON object.
-
-
-        Returns
-        -------
-        Object
-            The message as an object.
-
-        """
-
-        return dict({
-            "text": self.format_text(),
-            "files": self.get_files(),
-            "performers": self.get_performers(),
-            "price": self.get_price(),
-            "expiration": self.get_expiration(),
-            "schedule": self.get_schedule(),
-            "poll": self.get_poll(),
-            "keywords": self.get_keywords()
-        })
-
-    def get_schedule(self):
-        """
-        Gets the schedule value if not none else sets it from args or prompts.
-
-        Returns
-        -------
-        Schedule
-            Schedule object with proper values.
-
-        """
-
-        if self.schedule: return self.schedule
-        self.schedule = Schedule()
-        return self.schedule
+    @staticmethod
+    def format_poll(**pollArgs):
+        return Poll(**pollArgs).dump()
 
     def send(self):
         """
@@ -488,11 +246,10 @@ class Post(Message):
             Whether or not sending the post was successful.
 
         """
-        
-        self.init()
-        Settings.print("post > {}".format(self.get_text()))
-        if not self.get_files() and self.get_text() == "":
+
+        Settings.print("Posting...")
+        if not self.files and not self.text:
             Settings.err_print("Missing files and text!")
             return False
-        return post(self.get_post())
+        return WEBDRIVER_post(self.dump())