+++ OnlySnarf/classes/schedule.py
@@ -9,9 +9,9 @@ class ScheduleSchema(Schema):
     time = fields.Str()
     hour = fields.Str(default="00")
     minute = fields.Str(default="00")
-    year = fields.Int(default=0)
-    month = fields.Int(default=0)
-    day = fields.Int(default=0)
+    year = fields.Str(default="0")
+    month = fields.Str(default="0")
+    day = fields.Str(default="0")
     suffix = fields.Str(default="am")
 
     @post_load
@@ -20,18 +20,29 @@ class ScheduleSchema(Schema):
 
 class Schedule:
 
-    def __init__(self, schedule, date, time, hour, minute, year, month, day, suffix):
-        self._initialized_ = False
-        self.schedule = schedule
-        self.date = date
-        self.time = time
-        self.hour = hour
-        self.minute = minute
-        self.year = year
-        self.month = month
-        self.day = day
-        self.suffix = suffix
-        # self.init()
+    def __init__(self, date, time):
+        self.date = Schedule.format_date(date)
+        self.time = Schedule.format_time(time)
+
+
+        # TODO: finish updating schedule parsing
+        self.schedule = Schedule.format_schedule(schedule)
+
+
+        date = datetime.strptime(str(self.schedule), SCHEDULE_FORMAT)
+        self.year = date.year
+        self.month = date.strftime("%B")
+        self.day = date.day
+        self.hour = date.hour
+        self.minute = date.minute
+
+        self.suffix = "am"
+        if int(self.hour) > 12:
+            self.suffix = "pm"
+            self.hour = int(self.hour) - 12
+
+
+
 
     @staticmethod
     def create_schedule(schedule_data):
@@ -39,124 +50,15 @@ class Schedule:
         return schema.load(schedule_data)
 
     def dump(self):
+        if not self.validate(): return {}
         schema = ScheduleSchema()
         result = schema.dump(self)
         # pprint(result, indent=2)
         return result
 
-    def init(self):
-        """Initialize the schedule's settings"""
-
-        if self._initialized_: return
-        Settings.dev_print("initiliazing schedule...")
-        self.schedule = Settings.get_schedule()
-        date = datetime.strptime(str(self.schedule), SCHEDULE_FORMAT)
-        self.year = date.year
-        self.month = date.strftime("%B")
-        self.day = date.day
-        self.hour = date.hour
-        self.minute = date.minute
-        self.suffix = "am"
-        if int(self.hour) > 12:
-            self.suffix = "pm"
-            self.hour = int(self.hour) - 12
-        Settings.dev_print("year: {}".format(self.year))
-        Settings.dev_print("month: {}".format(self.month))
-        Settings.dev_print("day: {}".format(self.day))
-        Settings.dev_print("hour: {}".format(self.hour))
-        Settings.dev_print("minutes: {}".format(self.minute))
-        Settings.dev_print("suffix: {}".format(self.suffix))
-        Settings.dev_print("initiliazed schedule")
-        self._initialized_ = True
-
-    def get(self):
-        """
-        Get the schedule's values in a dict.
-
-        Returns
-        -------
-        dict
-            A dict containing the values of the schedule
-
-        """
-
-        return dict({
-            "date": self.get_date(),
-            "time": self.get_time(),
-            "hour" : self.hour,
-            "minute" : self.minute,
-            "year" : self.year,
-            "month" : self.month,
-            "day" : self.day,
-            "suffix" : self.suffix
-        })
-
-    def get_date(self):
-        """
-        Gets the date value if not none else sets it from args or prompts.
-
-        Returns
-        -------
-        str
-            The date as a valid date string
-
-        """
-
-        if self.date: return self.date
-        self.date = Settings.get_date()
-        return self.date
-
-        # prompt skip
-        # confirm date
-        if not Settings.confirm(date): return self.get_date()
-        self.date = date
-        return self.date
-
+    # TODO: return parsed schedule from self.date & self.time
     def get_schedule(self):
-        if self.schedule: return self.schedule
-        self.schedule = Settings.get_schedule()
-        return self.schedule
-
-    def get_time(self):
-        """
-        Gets the time value if not none else sets it from args or prompts.
-
-        Returns
-        -------
-        str
-            The time as a valid time string
-
-        """
-
-        if self.time: return self.time
-        # retrieve from args and return if exists
-        self.time = Settings.get_time()
-        return self.time
-
-
-
-        # # if time: 
-        # time = datetime.strptime(str(time), SCHEDULE_FORMAT)
-        # # Settings.dev_print(time)
-        # time = time.strftime("%I:%M %p")
-        # # Settings.dev_print(time)
-        # self.time = time
-        # return self.time
-
-        # retrieve time from schedule args and return if exists
-        schedule = Settings.get_schedule() or None
-        if schedule:
-            time = datetime.strptime(str(schedule), SCHEDULE_FORMAT)
-            # Settings.dev_print(time)
-            time = time.strftime("%I:%M %p")
-            # Settings.dev_print(time)
-            self.time = time
-            return self.time
-        # prompt skip
-        # confirm time
-        if not Settings.confirm(time): return self.get_time()
-        self.time = time
-        return self.time
+        pass
 
     def validate(self):
         """
@@ -171,8 +73,7 @@ class Schedule:
 
         Settings.dev_print("validating schedule...")
         today = datetime.strptime(str(datetime.now().strftime(SCHEDULE_FORMAT)), SCHEDULE_FORMAT)
-        # schedule = datetime.strptime(str(Settings.get_schedule().now().strftime(SCHEDULE_FORMAT)), SCHEDULE_FORMAT)
-        schedule = Settings.get_schedule()
+        schedule = self.get_schedule()
         if not schedule: return False
         if isinstance(schedule, str):
             schedule = datetime.strptime(schedule, SCHEDULE_FORMAT)
@@ -197,6 +98,12 @@ class Schedule:
 
 
 
+        self.hour = hour
+        self.minute = minute
+        self.year = year
+        self.month = month
+        self.day = day
+        self.suffix = suffix
 
 
 
@@ -245,3 +152,4 @@ class Schedule:
             schedule = datetime.strptime("{} {}".format(Settings.get_date(), Settings.get_time()), SCHEDULE_FORMAT).strftime(SCHEDULE_FORMAT)
         Settings.maybe_print("schedule (settings): {}".format(schedule))
         return str(schedule)[:20] # must be less than 19 characters
+